+++ lib/seo/breadcrumb/build.ex
@@ -1,13 +1,13 @@
 defprotocol SEO.Breadcrumb.Build do
   @fallback_to_any true
   @moduledoc """
-  Implement `build/1` which receives your item and return a `SEO.Breadcrumb.List.t` or `nil`
+  Implement `build/2` which receives your item and conn and returns a `SEO.Breadcrumb.List.t` or `nil`
   """
 
-  @spec build(term) :: SEO.Breadcrumb.List.t() | nil
-  def build(list_of_items)
+  @spec build(term, Plug.Conn.t()) :: SEO.Breadcrumb.List.t() | nil
+  def build(list_of_items, conn)
 end
 
 defimpl SEO.Breadcrumb.Build, for: Any do
-  def build(item), do: SEO.Breadcrumb.List.build(item)
+  def build(item, _conn), do: SEO.Breadcrumb.List.build(item)
 end