+++ lib/seo/site/build.ex
@@ -1,13 +1,13 @@
 defprotocol SEO.Site.Build do
   @fallback_to_any true
   @moduledoc """
-  Implement `build/1` which receives your item and return a `SEO.Site.t` or `nil`
+  Implement `build/2` which receives your item and conn and returns a `SEO.Site.t` or `nil`
   """
 
-  @spec build(term) :: SEO.Site.t() | nil
-  def build(item)
+  @spec build(term, Plug.Conn.t()) :: SEO.Site.t() | nil
+  def build(item, conn)
 end
 
 defimpl SEO.Site.Build, for: Any do
-  def build(item), do: SEO.Site.build(item)
+  def build(item, _conn), do: SEO.Site.build(item)
 end