+++ lib/seo/twitter/build.ex
@@ -1,13 +1,13 @@
 defprotocol SEO.Twitter.Build do
   @fallback_to_any true
   @moduledoc """
-  Implement `build/1` which receives your item and return a `SEO.Twitter.t` or `nil`
+  Implement `build/2` which receives your item and conn and returns a `SEO.Twitter.t` or `nil`
   """
 
-  @spec build(term) :: SEO.Twitter.t() | nil
-  def build(item)
+  @spec build(term, Plug.Conn.t()) :: SEO.Twitter.t() | nil
+  def build(item, conn)
 end
 
 defimpl SEO.Twitter.Build, for: Any do
-  def build(item), do: SEO.Twitter.build(item)
+  def build(item, _conn), do: SEO.Twitter.build(item)
 end