+++ lib/seo/unfurl.ex
@@ -67,7 +67,3 @@ defmodule SEO.Unfurl do
     """
   end
 end
-
-defimpl SEO.Unfurl.Build, for: Any do
-  def build(item), do: SEO.Unfurl.build(item)
-end