+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## 0.1.5 (October 17, 2022)
+
+- Allow for config to reference a module to call `config/0` and `config/1` upon
+- Allow for domain config to be a function reference that can receive a conn
+- BREAKING: rework implementation to be arity 2 that accepts the item and the conn
+
 ## 0.1.4 (October 16, 2022)
 
 - Fix empty breadcrumbs